backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
sqlalchemy>=2.0.0
alembic>=1.11.0
psycopg2-binary>=2.9.0
python-jose[cryptography]>=3.3.0
passlib==1.7.4
bcrypt==3.2.2
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
python-dateutil>=2.8.2
email-validator
